+++ bin/lib/contexts/ExecutionContext.js
@@ -193,4 +193,4 @@ const errors = {
     cannotAssignToConst: (c) => `cannot assign a value to a constant ${c}`,
     cannotAssignToOuterScope: (v) => `cannot assign a value to an outer scope variable ${v}`,
 };
-//# sourceMappingURL=Context.js.map
+//# sourceMappingURL=ExecutionContext.js.map
