+++ bin/lib/contexts/LoopContext.js
@@ -1,16 +1,16 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-const Context_1 = require("./Context");
+const ExecutionContext_1 = require("./ExecutionContext");
 const utils_1 = require("../utils");
 // CLASS DEFINITION
 // ================================================================================================
-class LoopContext extends Context_1.ExecutionContext {
+class LoopContext extends ExecutionContext_1.ExecutionContext {
     // CONSTRUCTOR
     // --------------------------------------------------------------------------------------------
     constructor(parent, inputs) {
         super(parent, undefined, inputs);
         this.blockResults = [];
-        this.rank = (parent instanceof Context_1.ExecutionContext ? parent.rank + 1 : 0);
+        this.rank = (parent instanceof ExecutionContext_1.ExecutionContext ? parent.rank + 1 : 0);
     }
     // ACCESSORS
     // --------------------------------------------------------------------------------------------
@@ -31,9 +31,38 @@ class LoopContext extends Context_1.ExecutionContext {
     setInputs(inputs) {
         // TODO: implement
     }
-    addBlock(blockResult) {
-        // TODO: validate blockResult expression
-        this.blockResults.push(blockResult);
+    addBaseBlock(initResult, segmentResults) {
+        // initializer result
+        const controller = this.getLoopController(this.rank);
+        let result = this.base.buildBinaryOperation('mul', initResult, controller);
+        // segment results
+        segmentResults.forEach((expression, i) => {
+            const resultControl = this.getSegmentController(i);
+            expression = this.base.buildBinaryOperation('mul', expression, resultControl);
+            result = this.base.buildBinaryOperation('add', result, expression);
+        });
+        // store result in a local variable
+        const resultHandle = `${this.id}t`; // TODO?
+        this.base.addLocal(initResult.dimensions, resultHandle); // TODO: better way to get dimensions
+        this.statements.push(this.base.buildStoreOperation(resultHandle, result));
+        this.blockResults.push(this.base.buildLoadExpression(`load.local`, resultHandle));
+    }
+    addLoopBlock(initResult, loopResult) {
+        // TODO: validate dimensions
+        const id = this.getLoopControllerId();
+        // initializer result
+        const controller = this.getLoopController(this.rank);
+        initResult = this.base.buildBinaryOperation('mul', initResult, controller);
+        // loop result
+        const one = this.base.buildLiteralValue(this.base.field.one);
+        const invController = this.base.buildBinaryOperation('sub', one, controller);
+        loopResult = this.base.buildBinaryOperation('mul', loopResult, invController);
+        // combine results and store them in a local variable
+        const resultHandle = `${this.id}t`; // TODO?
+        this.base.addLocal(initResult.dimensions, resultHandle); // TODO: better way to get dimensions
+        const result = this.base.buildBinaryOperation('add', initResult, loopResult);
+        this.statements.push(this.base.buildStoreOperation(resultHandle, result));
+        this.blockResults.push(this.base.buildLoadExpression(`load.local`, resultHandle));
     }
 }
 exports.LoopContext = LoopContext;
