+++ bin/lib/contexts/index.js
@@ -1,13 +1,9 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-var Context_1 = require("./Context");
-exports.ExecutionContext = Context_1.ExecutionContext;
+var ExecutionContext_1 = require("./ExecutionContext");
+exports.ExecutionContext = ExecutionContext_1.ExecutionContext;
 var LoopContext_1 = require("./LoopContext");
 exports.LoopContext = LoopContext_1.LoopContext;
-var LoopBlockContext_1 = require("./LoopBlockContext");
-exports.LoopBlockContext = LoopBlockContext_1.LoopBlockContext;
-var LoopBaseContext_1 = require("./LoopBaseContext");
-exports.LoopBaseContext = LoopBaseContext_1.LoopBaseContext;
 var RootContext_1 = require("./RootContext");
 exports.RootContext = RootContext_1.RootContext;
 //# sourceMappingURL=index.js.map
