+++ lib/contexts/index.ts
@@ -1,5 +1,3 @@
-export { Context, ExecutionContext } from './Context';
+export { Context, ExecutionContext } from './ExecutionContext';
 export { LoopContext } from './LoopContext';
-export { LoopBlockContext } from './LoopBlockContext';
-export { LoopBaseContext } from './LoopBaseContext';
 export { RootContext } from './RootContext';
